+++ README.md
@@ -1,2 +1,7 @@
 # ProtractorWorks
 Learning code for protractor with simple examples
+Notes on core:
+1) Keep all page files in 'page' folder.
+2) Page files should be named ending with '_po.js'. Eg: home_po.js
+3) Keep corresponding locators for each page in 'resources/locators' folder. Ex: home.json
+    a) Use the same name as that of page.
